Data Analysis-AHC and OOP
Performed AHC
In our OOP implementation, Node is a class. Each node in the AHC tree is
an object, and it has the following four variables at least:
• cluster: this variable saves the datapoints in the cluster as a list;
• height: this variable saves the height of the node in the AHC tree;
• leftChild: this variable saves the left child of this node; if the node is a
leaf and it has no left child, leftChild is set to None;
• rightChild: this variable saves the right child of this node; if the node is
a leaf and it has no right child, rightChild is set to None.
Certain parameters to note
-Generating Synthetic data
-Distance function
-Node,class and initiation of AHC
-Linkage functions
-Deciding which 2 clusters to merge
-AHC algorithm
-Cut the AHC and plot the datapoints in 3D