Machine Learning
For the description of the version of this dataset available in scikit-learn, run
print(diabetes.DESCR)
(as usual), where diabetes is my name for the dataset (you may use different
names, but please make sure they are easy to remember). The scikit-learn
function for loading this version of the dataset is load_diabetes.
For the main part of this assignment, you should use the original version of
the dataset, which is given here (the file is called “Tab-delimited diabetes data
(text file)”) and which requires preprocessing. (Alternatively, you can download
it from the module’s Moodle page.)
You should perform the following steps:
1. Load the scikit-learn version of the diabetes dataset into your Jupyter
notebook using the load_diabetes function.
2. Split the dataset into the training and test sets. You may use the function
train_test_split in scikit-learn. Here and below use your birthday
(in the format DDMM omitting leading zeros if any) as random_state.
3. What is the training and test R2
for the Lasso model using the default
parameters? How many features does this model use? What are the names
of those features? Write the answers in your Jupyter notebook. Here and
below, you are allowed to use any scikit-learn functions. (You may also
use numpy and matplotlib functions.)
4. Now load the original diabetes dataset from the web page given above.
Download the file
Tab-delimited diabetes data (text file)
by right-clicking on it. All the remaining tasks should be performed using
this file (diabetes.data), which is the original diabetes dataset. The
labels are given in the last column of the file diabetes.data.
5. Split the dataset into the training and test sets. Use your birthday (in the
format DDMM) as random_state.
6. Repeat item 3 for the current dataset. Comment on the differences from
what you saw in item 3.
7. Preprocess the training and test sets in the same way and avoiding data
snooping. Use StandardScaler.
8. Repeat item 3 for the current training and test sets (which you should
use in items 8–10). Are your current results closer to those in item 3 or
item 6? Notice that a priori you would expect your current results to be
closer to those in item 3, since the reason for different results in items 3
and 6 was that the former were for normalized data while the latter were
for the original data. Is this expectation confirmed? If not, why?
9. Varying the regularization parameter α in the Lasso, plot the test R2 vs
the number of features used (i.e., those with non-zero coefficients). Try to
make your plot as pretty as possible. (Obviously, it’s subjective.) Which
point on the curve do you prefer? (There is no unique correct answer to
this question.) Give a brief explanation of your preference.
10. Choose the regularization parameter for the Lasso using cross-validation
on the training set. Train the Lasso on the whole training set using the
chosen values of the parameters. Report the resulting training and test
R2 and the number of features used. (As before, you are allowed to use
any scikit-learn functions.)
11. Implement an inductive conformal predictor as follows:
(a) Split the training set that you obtained in item 5 into two parts:
the calibration set of size 99 and the rest of the training set (the
training set proper). Use your birthday (in the format DDMM) as
random_state.
(b) Preprocess the training set proper, calibration set, and test set in
the same way using StandardScaler. Namely, fit the scaler to the
training set proper and then use it to transform all three.
(c) Using the nonconformity measure α = |y − yˆ|, where y is the true
label and ˆy is its prediction given the training set proper, for each test
sample compute the prediction interval for it. Do this for significance
levels 5% and 20%. For each of these significance levels compute:
 the length of the prediction intervals for the test samples
 and the test error rate of your inductive conformal predictor1
.
For computing the predictions ˆy, use the Lasso with parameters chosen by cross-validation on the training set proper.
All results should be given in your Jupyter notebook (see above for a description
of what to include). Make sure to include the following numbers:
(a) The training and test R2
for the Lasso model with default parameters on
the scikit-learn version of diabetes and the number of features used.
(b) The training and test R2
for the Lasso model with default parameters on
the original version of diabetes and the number of features used.
(c) The training and test R2
for the Lasso model with default parameters on
your version of diabetes (i.e., on the original version of diabetes with
the features normalized by you); the number of features used.
(d) The training and test R2
for the Lasso model with the best parameters
chosen by cross-validation on your version of diabetes; the number of
features used.
(e) The lengths of prediction intervals and their test error rates at significance levels 5% and 20% (if you are implementing an inductive conformal
predictor).
There should be 16 numbers overall, plus your explanations (including the names
of chosen features) and a plot.
Inductive conformal prediction is described in Chapter 6. You are not allowed to use any existing implementations of inductive conformal prediction,
but you are allowed to use the functions available in scikit-learn, such as
Lasso, and NumPy, such as sort.