Machine learning
Be able to use and implement machine-learning algorithms, with the neural
networks, SVM, and cross-conformal prediction algorithms as examples. Have
an understanding of ways to apply the ideas and algorithms of machine learning
in industry.
Datasets
This assignment uses two datasets, wine (available in scikit-learn) and the
USPS dataset of handwritten digits. The latter dataset can be downloaded
from the web site of Hastie et al. (2009):
 Trevor Hastie, Robert Tibshirani, and Jerome Friedman. Web site for The
Elements of Statistical Learning: Data Mining, Inference, and Prediction,
Second Edition, https://web.stanford.edu/~hastie/ElemStatLearn/.
Click on the Data button on the left and go to the end (ZIP code). You will
need to guess the format of the dataset; since these are handwritten digits, the
labels are digits 0–9; the other entries are the pixels of their images. (Each
sample is represented by a line in the file; this single line is very long and will
be split into several on your computer screen.)
Tasks
Do the following for both datasets (you might prefer to start with the smaller
wine dataset):
1. Load the data set into Python using, e.g., load_wine or genfromtxt, as
appropriate. In the case of the USPS dataset, merge the original training
and test sets into one dataset.
2. Divide the dataset into a training set and a test set. You may use the
function train_test_split. Use your birthday in the format DDMM as
random_state (omit leading zeros if any).
3. Using cross-validation and the training set only, estimate the generalization accuracy of the MLPClassifier with the default values of the parameters. You may use the function cross_val_score.
4. Get rid of the warnings by varying a parameter, or parameters, such as
max_iter. Estimate the generalization accuracy of the MLPClassifier
with the modified values of the parameters. (If there are no warnings, the
modified values of the parameters may be identical to the default values.)
5. Find the test error rate of the MLPClassifier with the modified values
of parameters, compare it with the estimate obtained in the previous task
(task 4) for the same values of parameters, and write your observations in
a markdown cell of your Jupyter notebook.
6. Create a pipeline for MLPClassifier involving data normalization and
MLPClassifier, and use grid search and cross-validation to tune parameter alpha for the pipeline, avoiding data snooping and data leakage. You may use the scikit-learn class GridSearchCV. Experiment with different ways of doing normalization (such as StandardScaler, MinMaxScaler,
RobustScaler, and Normalizer). Which ways are appropriate for either
dataset? (The answer, which should be written in your Jupyter notebook,
may depend on the results that you obtain for the next task.)
7. Fit the GridSearchCV object of task 6 to the training set and use it to
predict the test labels. Write the resulting test error rate in your Jupyter
notebook.
8. Implement a cross-conformal predictor. You may use the KFold class for
splitting into folds (start from 5 or 10 folds). For computing the conformity
scores for each fold, you may use one of the GridSearchCV objects that
you created in task 6 in combination with the predict_proba method (see
Section 4 of Lab Worksheet 9 for an example). Run your cross-conformal
predictor on the two datasets, training it on the training set and testing
on the test set.
 To check its validity, produce a calibration curve, plotting the percentage of errors made on the test set1 vs the significance level
ϵ ∈ [0, 1].
 Compute the average false p-value on the test set.
9. An alternative to implementing a cross-conformal predictor is to experiment with a Support Vector Machine. Perform tasks 4–7 for the class
SVC in scikit-learn, as described in Lab Worksheet 8, Section 3. The
parameters to fit using grid search are C and gamma.
Some tasks might be hard to separate (such as 6 and 7), and it’s OK to implement several tasks in the same code cell.
Details
The cross-conformal predictor is described in Chapter 9. For other useful information and code, see Lab Worksheet 9 (e.g., the class KFold is described in
Section 2 and plotting calibration curves is discussed in Section 5).
You are not allowed to use any existing implementations of the cross-conformal
and related predictors, but you are allowed to use scikit-learn functions.
Make sure to include the following in your Jupyter notebook (for both
datasets) if you are implementing a cross-conformal predictor:
 Your estimate of the generalization accuracy of MLPClassifier with your
modified values of the parameters (task 4) and the test error rate for the
MLPClassifier with the same values of the parameters (task 5).
 The test error rate of your GridSearchCV object in task 7 for at least two
different ways of normalization (including Normalizer).
 The average false p-value and calibration plot for your cross-conformal
predictor.
Therefore, there should be at least 2 × (2 + 2 + 1) = 10 numbers and two plots
in your Jupyter notebook.
If you are experimenting with a Support Vector Machine, you should include
the following numbers (for both datasets and both classifiers):
 Your estimate of the generalization accuracy with your modified values of
the parameters (task 4) and the test error rate with the modified values
of the parameters (task 5).
 The test error rate of your GridSearchCV object in task 7 for at least two
different ways of normalization (including Normalizer).
Therefore, there should be at least 2 × 2 × (2 + 2) = 16 numbers. Remember
that the modified values of the parameters may coincide with the default values
if there are no warnings.